11-04: replacement cycles by archetype, and the two identities
`archetypes.<name>.d.<good>` multiplies a good's life, so a type differs in
how often it turns up at the shelf and not only in what it will pay when it
does. `flow_cost` divides by the household's own life, `Flow.Spread` takes a
real span, and a life is whatever `life_g · d[A][g]` comes to — 20.27 months
is a life a geometric distribution can have.

**Two identities, conserving two different things.** `Σ share · m = 1` fixes
the score level so §3.6's base scores keep meaning what they say at the
population mean. `Σ share / d = 1` fixes the units so `capacity =
round(households / life_g)` stays correct as written. Neither implies the
other and a table can satisfy one while breaking the other.

**The second is harmonic, and the loader multiplies where it divides for `w`.**
Demand per tick is `1/life`, so it is the reciprocal that has to average to
one. Normalise `d` itself and Jensen hands the population strictly more
replacement demand than capacity was sized for — permanently, in every good,
in every run, with the reprice rule absorbing it into the price level and
every conservation identity still holding to the cent. The story asked for
the test before the normalisation; the claim that nothing else catches it was
checked rather than assumed. With the normalisation turned arithmetic, **587
of 593 tests still pass** and all six failures are in ReplacementCycleTests.
The catching test rescales §3.7's phone column so its arithmetic mean is 1
and measures the 6.7% overshoot in the direction Jensen guarantees.

**`ŵ = m / d` is derived, not authored.** Because the cost divides by the
household's own life, a type that replaces its phone 48% more often has
already spent almost all of the 1.55 weight meant to make it the top phone
bidder: author `ŵ` directly and `gadget` comes out at 1.050 against
`prudent`'s 1.087, bidding below the type defined by keeping its phone. The
composition is measured through the engine rather than argued — a `gadget`
household's realised score multiplier on a phone is 1.553, its authored `m`,
recovered from `Taste` and `Life` on a drawn population.

**Taste is keyed by category label, the cycle by good.** A type wants
electronics; it churns a phone and keeps a television, and §3.7 has three
different cycles inside electronics. Under §3.1 a label is a row and the
distinction is invisible, which is why every file written before E11 still
means what it meant. The loader rejects a good's name in a `w` and a label in
a `d`.

A `d` is rejected, never rounded, wherever it cannot mean what it says:
under `deterministic` (deferred here from 11-03, which had no `d` to reject),
on a life-1 good, and where the normalised table would put a realised life
below one tick — at which point `1/life` stops being a probability and a
durable quietly becomes a consumable at a durable's price. Reported per good
rather than per (type, good), because normalisation is per column and one
stated cycle scales every other type's 1.0 off 1 as well.

`capacity` stays `round(households / life_g)` and is **not** derived from the
realised assignment: two seeds of one scenario produce identical effective
configurations, which is what the campaign collector refuses to proceed
without. The finite-sample residue is reported once in `run.done` and left
alone. Reporting it found a third source §3.7 never named and the largest of
the three on a narrow shelf: `round` itself. At 1,000 households §3.1's
appliances are `1000/96 = 10.4` units rounded to 10, so that shelf is 4.2%
undersized before an archetype is drawn — most of the argument for §3.6's
bigger town, and none of an argument for a cleverer capacity rule.

`d ≡ 1` reproduces byte for byte, on the same snapping `Scale` V5a rests on.
V5's baselines are regenerated: every CSV is byte-identical across all four
seeds, and only the effective config (one `d` line) and the marker (two keys)
moved.

Also corrects §3.7's claim that the hazard is quieter than the calendar,
which 11-03 disproved in `01-SIMULATION.md` §5.5 but left standing here.

594 tests; V2, V3, V4, V4 under the hazard, V5 and V5a green.

+++ src/EconomySimulation.Engine/Decision/Candidate.cs
@@ -14,14 +14,14 @@ namespace EconomySimulation.Engine.Decision;
 /// <param name="Tier">The tier the step lands on. Tier 0 is a purchase; anything above is an upgrade.</param>
 /// <param name="DeltaValue">What the step adds, in euros per tick: `V · (value_mult_t − value_mult_(t−1))`.</param>
 /// <param name="DeltaPrice">What the step costs in cash, now: `price_t − price_(t−1)`. The increments telescope to the tier's posted price exactly.</param>
-/// <param name="Life">The category's life, which turns the cash increment into a per-tick cost.</param>
+/// <param name="Life">The life **this household** gets out of the good, which turns the cash increment into a per-tick cost. `life_g · d[A][g]` under §3.7, and the good's own life wherever `d` is 1.</param>
 /// <param name="Score">`Δvalue / Δcost`, a pure number, compared against λ.</param>
 public readonly record struct Candidate(
     int Category,
     int Tier,
     Flow DeltaValue,
     Money DeltaPrice,
-    int Life,
+    double Life,
     double Score)
 {
     /// <summary>An upgrade needs the step below it taken first; a purchase needs nothing.</summary>
